Add view to confirm tokens via Links in Emails.
This adds support to confirm requests sent to users by clicking on links which
some consider to be more user friendly than replying to emails with
confirmation tokens.

This URL (/list/<listid>/confirm/?token=<token>) would then need to be added to
the templates using $token as a variable, to be filled in by Mailman Core when
sending out email.