Add support for viewing list-options for multiple user subs.
When a User is subscribed to the same list with two addresses, they can only
look at a single membership's preferences since the URL only has a list id
componenet.

This MR adds another component, member_id, to the list options page which can
individually refer to a specific membership. At this point the list_id is
actually redundat and can be removed since the Member ties both a list to a
specific subscriber.

I am not sure if replacing the list-id with member_id makes sense, but it
possible to so.