Fix the change subscription process.
When trying to change a subscripton from an address that was subscribed to
primary address (which happens to be the same old address), we'd fail with an
AssertionError as we'd fail to find the Member resource.

Add a new member_id field to change subscription form making it explicit about
the Membership we want to change the address on.

Also, handle a corner case when the case whe the destination address or user_id
is already subscribed.