Remove signal handler to create a Mailman User and some refactoring.
Since django-mailman3 will now always create a new mailman user when a User
object is created, we don't want to duplicate the signal handlers.

Also, we don't need the duplicate methods in MailmanUserManager to do the same
job as get_mailman_user() which has same semantics and get_or_create and is
used at more places.