chore: Update test setup such that it runs from project root.
Currently, the pytest-django is setup in a way that example_project is the
root for the testing purposes and we specify paths like `../src` to the
commands that need to run only _some_ tests compared to all. We also
have to do things weirdly for docs and other kinds of commands that need to
read Django settings etc.

This change makes it such that we use project root as the reference directory
when writing commands. This also modifies the appropriate commands in tox.ini
to adapt to this change.

Update documentation to reflect how the tests need to be run for specific tox
environment and for specific test files.